backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
sqlalchemy==1.4.26
alembic==1.7.5
mysql-connector-python==8.0.26
python-multipart==0.0.5
requests==2.26.0
mutagen==1.45.1
guessit==3.4.0
pysrt==1.1.2
webvtt-py==0.4.5
ffmpeg-python==0.2.0
chardet==4.0.0